+++ example/src/main/AndroidManifest.xml
@@ -45,8 +45,8 @@
             android:label="@string/activity_title_demo_d_grid"/>
 
         <activity
-            android:name=".demo_d_with_longpress.DraggableWithLongPressExampleActivity"
-            android:label="@string/activity_title_demo_d_with_longpress"/>
+            android:name=".demo_d_on_longpress.DragOnLongPressExampleActivity"
+            android:label="@string/activity_title_demo_d_on_longpress"/>
 
         <activity
             android:name=".demo_s.SwipeableExampleActivity"
